bulk commit (not finished)
